+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog for NVDB API LES V3 Client
+
 ## 1.6.0
 * `Statistics.length` changed from `long` to `double`
+* `ClientConfiguration` has been added to allow setting read and connect timeouts for Jersey client.
 
 ## 1.5.0
 * RoadNetRequest.id, .superId changed from `List<Integer>` to `List<Long>`